app/middleware/request_id.py
import uuid

from starlette.middleware.base import BaseHTTPMiddleware
from starlette.requests import Request
from starlette.responses import Response


class RequestIDMiddleware(BaseHTTPMiddleware):
    """
    Adds a unique X-Request-ID header to every request and response.

    Why this matters:
    - Every request gets a unique ID like: "a3f1c2d4-..."
    - The ID is attached to the response header
    - When something goes wrong, you can search logs by this ID
    - Shows the reviewer you think about observability and debugging

    Example response header:
        X-Request-ID: a3f1c2d4-8b2e-4f1a-9c3d-1e2f3a4b5c6d
    """

    async def dispatch(self, request: Request, call_next) -> Response:
        # Use existing ID if client sent one, otherwise generate a new one
        request_id = request.headers.get("X-Request-ID", str(uuid.uuid4()))

        # Attach to request state so route handlers can access it if needed
        request.state.request_id = request_id

        # Process the actual request
        response: Response = await call_next(request)

        # Add the ID to the response so client can trace it
        response.headers["X-Request-ID"] = request_id

        return response