requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
 
# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
 
# Data validation
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
 
# Authentication
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-multipart==0.0.12
 
# Groq AI
groq==0.11.0
 
# Rate limiting
slowapi==0.1.9
 
# Testing
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0